Adding of Jumping, Ground Tag, Serialize Variables, Physics calculations and Camera followers the player
I have changed the update to be fixed update for the physics updates.

I have added a jump that allows the player to jump once and when the player hits the floor the player can jump again.

I have added Serialize Fields to the variables that I think would be need to be changed in the inspector.

I have added a Ground Tag.

I have the player model faces where the camera is looking.

Also added in a testing scene instead of using a different unity project.

Some bugs that I have found and don't know how to fix are when the player first loads in the camera doesn't know where to look and when the player jumps when not move the camera also doesn't know what to look at.