Remade the player movement script
The old movement script component have been deleted from the project but I have created backups of the code just in case the new code is not wanted.

The camera has been made a child of the player to get around the camera not knowing where it is need to be when jumping or moving. The cursor has been locked in the centre of the screen.

A sprint button has been implemented into the script and has been set to left shift

The player is check when jumping that it is on the ground using a character controller. when falling the player slowly increase the fall speed.

Time.fixedDeltaTime has been used in the fixed update instead of time.DeltaTime.