change to rigidbody, camera code separated
The character controller code has been taken out and replace with a rigidbody.

The code for the camera movement and rotation have also been removed and put in a different script.

Removed some of the comments of the code. However let me know if it is still over commented.

The player could have got higher jump height when chaining jumps together that has been fixed.