Removing tertinary operators and change to forces
The tertinary operators have been removed and replaced with a if statement that makes the code more readable.

Move.Position has been removed and replaced with forces and it does acceleration.

The locked cursor and has been put in separate method

The player can't move forward and left or right to get more speed.

When I was testing the camera moving left to right it was working. If the problem is still occurring can you show me it not working.