Implementing the feedback from the code review & Light Flickering
I have used fixeddeltatime on the physics calculations.

The camera now teleports to the player on every frame.

The Character Movement script use Input.GetAxisRaw instead of GetAxis.

An empty object has been made and the player and camera are child of this game object.

IsGrounded has become its own method

The rb.adddrag has been used to put drag on the player and gets set this drag to 0 when not grounded.

I have also created a simple light flickering system that has a changeable intensity and flicker speed. It change the intensity of the light over time.