removed magic numbers, Player's actual speed variable
I have removed the magic numbers from the drag method and replaced them with variables that can be changed in the inspector.

The Current Speed variable get the actual player speed from the rigid body instead of the change of states between the Move Speed and Sprint Speed. The Current Speed variable is also public.

A new variable was made to take the role of Current Speed so it doesn't mess up the actual speed.