Player Slope Movement and Fixed Jump
The player can now walk up slopes and stairs with a slope over it.

Fixed the issue with the jump where you had to press space on every frame. you can now press the spacebar and it should jump also fixed a bug where the player could bhop around the map.