Working item pick up system
Clicking F allows an item to be picked up.
This is done using my own movement system as I couldn't figure out the one made for the game.
When it is needed to be put into the player prefab, all that needs to be done is the script to be put into the *player camera* and an empty child under the camera to be the position in which items are held when picked up.