Library_Stairs script fully implemented?
For some reason the "Lever" I set up doesn't change state when interacted with? It had the 'Interactable Switch' Script and I was using the player prefab which should be able to interact with it, but...it didn't work, at least in my testing level. I assume this script works and I have tested it - though exclusively in the 'false' state due to prior limitations.