Updates to the player movement & Camera
I have create two scripts that are in the temp folder under scripts in my name. The player Prefab has been changed so it uses the new scripts for the movement  and camera should feel better and not lag on lower fps or jitter.

The code has been organise so it easier to read however I can added more comments to the code on bits of the code people don't understand.

The movement speed and jump force will need to be edit as it jumps pretty high.

Code that ben put into the other camera script has been moved over.