Library_Stairs Code Update
Stairs should now be able to function within negative values. The speed at which the stairs move might be slower than usual if the Start_Pos is a positive value and the End_Pos is a negative value, but that can be compensated for by adjusting the Speed variable accordingly.

Levers now use the Interaction system (I swear I did that before, but I can't find it.)